Created ignoreGameTitle list so categories that do not contains game names can be ignored, also created validPlatforms list so valid platform icons can be added.
Added a maxWidth of 300px to the hoverContainer.

Added iconUrls (xbox, playstation, pc, etc...) using chrome extension to get full URL of the icon resources.

Created updateGameInfo(), this is used to update hoverContainer game information based on retrieved gameData information. This allows the hover container to display, the games name, playable platforms, genres, price, and release date.

Created retrieveGameData(), this function retrieves the pages game title and sends it as a message to background.js to retrieve its appID and data.