Created gameData object to store all retrieved game information (release date, genre, name, websites, price, and platforms)
Created formatGameTitle(), used to properly format requested game titles for searching using API request

Changed request in getAppID function to host the request on a local server.

Created getGameID() to retrieve gameID of a request game from the igdb API.

Created getGameDetails() to store retrieved game details retrieved from steam API and igdb API.

Created checkResponse(), this was created to check the response from our requests, specifically of all data is present and structured properly.

Created getGameIDSrict(), this is used to retrieve the requested games id, by retrieving games that only strictly match the title of the requested game. This function runs twice if the response is empty, it searches for the formatted version of the games title and the original version if a response has not been retrieved. This stores candidates of gameIDs where responses match the games name and have a gameID within their response.

Created processGame(), this function retrieves all gameIDs and validates their responses.